Fix for ammo capacity blueprints
The blueprints for the ammo capacity increase of point defence,
heat sink launcher, and chaff launcher have been made more rational
in-game.  Instead of having 3 grades with only grade 3 giving an
extra usable shot, there is now only a grade 1 blueprint that results
in a capacity increase.  This blueprint has the cost of the old g3
blueprint.

Additionally, all the blueprints for these utilities are available
from both Ram Tah in the bubble and Petra Olmanova in Colonia.